+++ src/main/java/com/example/Sandwichi/services/ClientService.java
@@ -2,6 +2,7 @@ package com.example.Sandwichi.services;
 
 import com.example.Sandwichi.entities.Client;
 import com.example.Sandwichi.repositories.ClientRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -27,6 +28,10 @@ public class ClientService {
     }
 
     public void deleteClient(Long id) {
-        clientRepository.deleteById(id);
+        if (clientRepository.existsById(id)) {
+            clientRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Client not found with id " + id);
+        }
     }
 }
